refactor: move button and get/set to parent component #113
